+++ inc/session.class.php
@@ -47,7 +47,7 @@ class PluginDatainjectionSession
         if (!isset($_SESSION['datainjection'][$param])) {
             return false;
         }
-        if (in_array($param, ['results', 'error_lines'])) {
+        if (in_array($param, ['results', 'error_lines', 'injection_lines', 'injection_results', 'injection_error_lines'])) {
             $fic = $_SESSION['datainjection'][$param];
             return file_get_contents(GLPI_TMP_DIR . '/' . $fic);
         }
@@ -66,7 +66,13 @@ class PluginDatainjectionSession
     public static function setParam($param, $results): void
     {
 
-        if (in_array($param, ['results', 'error_lines'])) {
+        if (in_array($param, ['results', 'error_lines', 'injection_lines', 'injection_results', 'injection_error_lines'])) {
+            if (isset($_SESSION['datainjection'][$param])) {
+                $old_fic = GLPI_TMP_DIR . '/' . $_SESSION['datainjection'][$param];
+                if (file_exists($old_fic)) {
+                    unlink($old_fic);
+                }
+            }
             $fic = Session::getLoginUserID() . '_' . $param . '_' . microtime(true);
             file_put_contents(GLPI_TMP_DIR . '/' . $fic, $results);
             $_SESSION['datainjection'][$param] = $fic;
@@ -84,11 +90,14 @@ class PluginDatainjectionSession
     public static function removeParams(): void
     {
 
-        if (isset($_SESSION['datainjection']['results'])) {
-            unlink(GLPI_TMP_DIR . '/' . $_SESSION['datainjection']['results']);
-        }
-        if (isset($_SESSION['datainjection']['error_lines'])) {
-            unlink(GLPI_TMP_DIR . '/' . $_SESSION['datainjection']['error_lines']);
+        $file_params = ['results', 'error_lines', 'injection_lines', 'injection_results', 'injection_error_lines'];
+        foreach ($file_params as $param) {
+            if (isset($_SESSION['datainjection'][$param])) {
+                $fic = GLPI_TMP_DIR . '/' . $_SESSION['datainjection'][$param];
+                if (file_exists($fic)) {
+                    unlink($fic);
+                }
+            }
         }
         unset($_SESSION['datainjection']);
     }