+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix responsible group injection payload normalization so group remains visible in GLPI after import
 - Fix incorrect escaping of apostrophes and accents
 - Fix plugin rights initialization and cleanup
+- Fix injection loading bar
 
 
 ## [2.15.4] - 2026-03-16